Add support for RSS / Atom extensions
